requirements.txt
python==3.10
faiss-cpu
sentence-transformers
deep-translator